+++ changelog.md
@@ -1,4 +1,15 @@
 # Changelog
 
-**./public/scripts/market-data.js - handleDataSort**
-* Removed commented out array reverse call.
+**Commented Files**
+* ./public/scripts/local-req.js
+* ./public/scripts/market-data.js
+* ./public/stylesheets/personal.css
+* ./views/index.ejs
+
+---
+
+**./public/scripts/market-data.js**
+* Variable declarations and assignments are now separate:
+	* 'alphaStatus' (handleDataSort)
+	* 'inputElement' (callSearch)
+	* 'tableBody' (renderCurrencyData)